Revise package metadata for permanent fork.
Sergey does not want to support streaming compression and
decompression in zstd, and that’s an unacceptable limitation relative
to the compression modules in the standard library, so unless he
changes his mind, it looks like this is going to be a permanent fork.

And as long as it’s going to be a permanent fork, we may as well call
it zstd 2.0 and clean out all the legacy gunk.