Remove compress_old and decompress_old and --legacy configuration.
Instead, decompress() can now handle both normal zstd frames and
frames with the pre-1.0.0.99 header, and there is now a
zstd.MIN_LEGACY_FORMAT constant that tells you the oldest format
version that can be decompressed.

In order to test all of this thoroughly enough for my tastes, it was
necessary to teach decompress() how to decompress frames for which
ZSTD_getFrameContentSize returns ZSTD_CONTENTSIZE_UNKNOWN, which is
nice because that was on the to-do list anyway.